Cef update && mouse for cef is ok.

+++ API/RDR2/Game.cs
@@ -176,15 +176,15 @@ namespace RDRN_API
 		}*/
 		public static bool IsControlPressed(int index, Control control)
 		{
-			return Function.Call<bool>(Hash.IS_DISABLED_CONTROL_PRESSED, index, (int)control);
+			return Function.Call<bool>(Hash.IS_DISABLED_CONTROL_PRESSED, index, control);
 		}
 		public static bool IsControlJustPressed(int index, Control control)
 		{
-			return Function.Call<bool>(Hash.IS_DISABLED_CONTROL_JUST_PRESSED, index, (int)control);
+			return Function.Call<bool>(Hash.IS_DISABLED_CONTROL_JUST_PRESSED, index, control);
 		}
 		public static bool IsControlJustReleased(int index, Control control)
 		{
-			return Function.Call<bool>(Hash.IS_DISABLED_CONTROL_JUST_RELEASED, index, (int)control);
+			return Function.Call<bool>(Hash.IS_DISABLED_CONTROL_JUST_RELEASED, index, control);
 		}
 		public static bool IsDisabledControlPressed(int index, Control control)
 		{
@@ -200,30 +200,30 @@ namespace RDRN_API
 		}
 		public static bool IsEnabledControlPressed(int index, Control control)
 		{
-			return Function.Call<bool>(Hash.IS_CONTROL_PRESSED, index, (int)control);
+			return Function.Call<bool>(Hash.IS_CONTROL_PRESSED, index, control);
 		}
 		public static bool IsEnabledControlJustPressed(int index, Control control)
 		{
-			return Function.Call<bool>(Hash.IS_CONTROL_JUST_PRESSED, index, (int)control);
+			return Function.Call<bool>(Hash.IS_CONTROL_JUST_PRESSED, index, control);
 		}
 		public static bool IsEnabledControlJustReleased(int index, Control control)
 		{
-			return Function.Call<bool>(Hash.IS_CONTROL_JUST_RELEASED, index, (int)control);
+			return Function.Call<bool>(Hash.IS_CONTROL_JUST_RELEASED, index, control);
 		}
 
 		public static bool IsControlEnabled(int index, Control control)
 		{
-			return Function.Call<bool>(Hash.IS_CONTROL_ENABLED, index, (int)control);
+			return Function.Call<bool>(Hash.IS_CONTROL_ENABLED, index, control);
 		}
 
 		public static void EnableControlThisFrame(int index, Control control)
 		{
-			Function.Call(Hash.ENABLE_CONTROL_ACTION, index, (int)control, true);
+			Function.Call(Hash.ENABLE_CONTROL_ACTION, index, control, true);
 		}
 
 		public static void DisableControlThisFrame(int index, Control control)
 		{
-			Function.Call(Hash.DISABLE_CONTROL_ACTION, index, (int)control, true);
+			Function.Call(Hash.DISABLE_CONTROL_ACTION, index, control, true);
 		}
 
 		public static void DisableAllControlsThisFrame(int index)
